Fix JWT TTL type casting in registration response
Cast config('jwt.ttl') to int to prevent Carbon TypeError when calculating
token expiration. The .env value was being read as string, causing
'Argument #3 ($value) must be of type int|float, string given' error.

+++ app/Http/Controllers/Api/V1/AuthController.php
@@ -79,7 +79,7 @@ class AuthController extends Controller
                         'user' => $user->load('tenant'),
                         'token' => $token,
                         'token_type' => 'bearer',
-                        'expires_in' => config('jwt.ttl') * 60, // Convert minutes to seconds
+                        'expires_in' => (int) config('jwt.ttl') * 60, // Convert minutes to seconds
                         'session_id' => $sessionId,
                     ],
                 ], 201);
